+++ src/main/java/com/ra/session05ex01/service/EnrollmentService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface EnrollmentService {
     List<Enrollment> getAllEnrollments();
     Enrollment getEnrollmentById(int id);
-    Enrollment createEnrollment(String studentName, int courseId);
-    Enrollment updateEnrollment(int id, String studentName, int courseId);
+    Enrollment createEnrollment(String studentName, Long courseId);
+    Enrollment updateEnrollment(int id, String studentName, Long courseId);
     Enrollment deleteEnrollmentById(int id);
 }
